added torch_dtype

+++ CALM-experiments/key_value_pairings_finetuning.py
@@ -151,7 +151,7 @@ training_args = TrainingArguments(
     do_predict=True,
 )
 
-model = AutoModelForCausalLM.from_pretrained(model_id, device_map="auto")
+model = AutoModelForCausalLM.from_pretrained(model_id, device_map="auto", torch_dtype=torch.bfloat16)
 device = torch.device("cuda")
 model.to(device)
